Cleanup AbstractUpdateMojo
(cherry picked from commit e6b7f5525f4c161f977e89b1590477607aad795b)
