+++ tycho-extras/tycho-version-bump-plugin/src/main/java/org/eclipse/tycho/versionbump/AbstractUpdateMojo.java
@@ -14,59 +14,47 @@
 package org.eclipse.tycho.versionbump;
 
 import java.io.File;
-import java.util.Collections;
 
+import org.apache.maven.execution.MavenSession;
 import org.apache.maven.plugin.AbstractMojo;
 import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.plugin.MojoFailureException;
 import org.apache.maven.plugins.annotations.Component;
 import org.apache.maven.plugins.annotations.Parameter;
 import org.apache.maven.project.MavenProject;
-import org.codehaus.plexus.logging.Logger;
-import org.eclipse.tycho.TargetEnvironment;
-import org.eclipse.tycho.core.resolver.P2Resolver;
-import org.eclipse.tycho.core.resolver.P2ResolverFactory;
 
 public abstract class AbstractUpdateMojo extends AbstractMojo {
-
-    @Component
-    private Logger logger;
-
-    @Parameter(defaultValue = "JavaSE-17")
-    private String executionEnvironment;
 
     @Parameter(property = "project")
     private MavenProject project;
 
     @Component
-    private P2ResolverFactory factory;
+    private MavenSession mavenSession;
+
+    MavenSession getMavenSession() {
+        return mavenSession;
+    }
 
     public void execute() throws MojoExecutionException, MojoFailureException {
-        try {
-            doUpdate();
-        } catch (Exception e) {
-            if (e instanceof MojoFailureException mfe) {
-                throw mfe;
+        File file = getFileToBeUpdated();
+        if (file != null) {
+            try {
+                doUpdate(file);
+            } catch (Exception e) {
+                if (e instanceof MojoFailureException mfe) {
+                    throw mfe;
+                }
+                if (e instanceof MojoExecutionException mee) {
+                    throw mee;
+                }
+                throw new MojoExecutionException("Could not update " + getFileToBeUpdated(), e);
             }
-            if (e instanceof MojoExecutionException mee) {
-                throw mee;
-            }
-            throw new MojoExecutionException("Could not update " + getFileToBeUpdated(), e);
         }
     }
 
     protected abstract File getFileToBeUpdated() throws MojoExecutionException, MojoFailureException;
 
-    protected abstract void doUpdate() throws Exception;
-
-    P2Resolver createResolver() {
-        return factory.createResolver(Collections.singletonList(TargetEnvironment.getRunningEnvironment()));
-
-    }
-
-    String getExecutionEnvironment() {
-        return executionEnvironment;
-    }
+    protected abstract void doUpdate(File file) throws Exception;
 
     MavenProject getProject() {
         return project;
